requirements.txt
cycler==0.10.0
Django==1.9.4
django-autocomplete-light==3.1.4
django-crispy-forms==1.6.0
django-registration==2.0.4
matplotlib==1.5.1
numpy==1.10.4
pandas==0.18.0
psycopg2==2.6.1
pyparsing==2.1.0
python-dateutil==2.5.0
pytz==2016.1
Quandl==2.8.9
requests==2.9.1
six==1.10.0
ystockquote==0.2.4
